requirements.txt
streamlit
numpy
Pillow
pandas
geopy
tensorflow